content/base/test/test_tor_bug15502.html

<!DOCTYPE HTML>
<html>
<!--
https://bugs.torproject.org/15502
-->
<head>
  <meta http-equiv="content-type" content="text/html; charset=utf-8">
  <title>Test for Tor Browser Bug 15502</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript;version=1.7" src="bug15502_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content"></div>

<script class="testbody" type="application/javascript;version=1.7">
SimpleTest.waitForExplicitFinish();

let prefs = SpecialPowers.Cu.import("resource://gre/modules/Services.jsm").Services.prefs;

let domain1 = "http://example.com",
    domain2 = "http://example.net",
    path = "/tests/content/base/test/",
    page_blob = "bug15502_page_blobify.html",
    page_deblob = "bug15502_page_deblobify.html"
    worker_blob = "bug15502_worker_blobify.html",
    worker_deblob = "bug15502_worker_deblobify.html";

let tabIO = function* (domain, child, input) {
  tab = window.open(domain + path + "bug15502_tab.html", "_blank");
  yield receiveMessage(tab); // ready message
  sendMessage(tab, "http://example.org" + path + child);
  yield receiveMessage(tab); // ready message
  sendMessage(tab, input);
  return yield receiveMessage(tab);
};

let blobTest = function* (domainA, domainB, blobPage, deblobPage, isolationOn) {
  prefs.setIntPref("privacy.thirdparty.isolate", isolationOn ? 2 : 0);
  let input = "" + Math.random();
  let blobURL = yield tabIO(domainA, blobPage, input);
  let result = yield tabIO(domainB, deblobPage, blobURL);
  let description = domainA + ":" + blobPage + "->" + domainB + ":" + deblobPage + ", isolation " + (isolationOn ? "on." : "off.")
  if (isolationOn && domainA !== domainB) {
    ok(input !== result, description + " Deny retrieval");
  } else {
    ok(input === result, description + " Allow retrieval");
  }
};

// "bug15502_page_deblobify.html",

Task.spawn(function* () {
  for (let isolate of [false, true]) {
    for (let domainB of [domain1, domain2]) {
      for (let blob of [page_blob, worker_blob]) {
        for (let deblob of [page_deblob, worker_deblob]) {
          yield blobTest(domain1, domainB, blob, deblob, isolate);
        }
      }
    }
  }
  SimpleTest.finish();
});

</script>

</body>
</html>
